oy. async validations with ember-changeset
validating the preferred username field requires a server round trip,
which leads to nasty UI hiccups when the user types a few characters at
a time. the field value, driven by the validated changeset value, will
"catch up" as the sequence of network calls complete, leading to a field
that "mysteriously" changes value as you type.

ember-changeset has an issue open to deal with this kind of thing. until
that lands or we decide to go in on ember-concurrency, this situation
is handled using an observer on a binding for the value which requires
async validation. this allows us to debounce setting the actual changeset
value, which keeps the remote validations to a sane rhythm.

the two gotchas are that we need to explicitly say the values are valid
if they haven't changed (in the remote validator) and then also manually
rollback the temporary binding.