refactor basic card save action
simplify the validations that are happening, since the server will do
its own validations anyway. validating preferredUsername at this point
was muddying up the reasonability of the thing. we just care if the
email needs to be verified so that we know whether to show the modal
first or not.